+++ server/index.js
@@ -1,17 +1,16 @@
-const express = require("express");
 const bodyParser = require("body-parser");
 const cookieParser = require("cookie-parser");
-
+const express = require("express");
 const expressWs = require("express-ws");
-const wsConnectHandler = require("./lib/webSocket").connect;
+const pino = require('express-pino-logger')();
 
 const usersHandlers = require("./routes/users");
 const problemsHandlers = require("./routes/problems");
 const publicationsHandlers = require("./routes/publications");
 const orcidAuthHandlers = require("./routes/auth/orcid");
+const feedbackHandlers = require("./routes/feedback");
 
-const fb = require("./feedback");
-
+const wsConnectHandler = require("./lib/webSocket").connect;
 const blobService = require("./lib/blobService");
 const upload = blobService.upload;
 blobService.initialise();
@@ -22,7 +21,7 @@ function noop () {}
 const app = express();
 const webSocketInstance = expressWs(app).getWss();
 
-const pino = require('express-pino-logger')();
+// attach pino logger
 app.use(pino);
 
 // terminate abandoned open sockets, 30s interval
@@ -46,28 +45,24 @@ global.subscriptions = new Map();
 // app.use(express.static("public"));
 
 app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
 app.use(cookieParser());
-app.use(
-  bodyParser.urlencoded({
-    extended: true,
-  }),
-);
+
+app.ws("/api", wsConnectHandler);
 
 app.get("/api", (req, res) => {
   res.send("Welcome to the Octopus API!");
 });
-app.ws("/api", wsConnectHandler);
 
 app.use("/api/problems", problemsHandlers.router);
 app.use("/api/publications", publicationsHandlers.router);
 app.use("/api/users", usersHandlers.router);
 app.use("/api/auth/orcid", orcidAuthHandlers.router);
 
-app.post("/api/feedback", fb.postFeedback);
-app.post(
-  "/api/image",
+app.post("/api/feedback", feedbackHandlers.postFeedback);
+app.post("/api/image",
   upload(blobService.AZURE_FEEDBACK_IMAGE_CONTAINER).single("image"),
-  fb.postImage,
+  feedbackHandlers.postImage,
 );
 
 app.use(function(err, req, res, next) {
